[calib] MAJOR provide explanation for detector behavior over time!
I did not expect this to be so obvious. :O Plotting the data against
the environment temperature was a bit of a "well, we can try it,
what's the worst that can happen" like idea. I almost didn't continue
when I only saw the 55Fe mean peaks and hadn't added the gas gains yet!